스키마 변경, 도메인 객체 타입 변경

+++ src/main/java/kr/co/hrmanager/domain/departments/Departments.java
@@ -12,7 +12,7 @@ import javax.persistence.*;
 public class Departments {
     @Id
 //    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer deptId;
+    private Long deptId;
 
     private String deptName;
