+++ src/main/java/kr/co/hrmanager/domain/employees/EmployeesRepository.java
@@ -2,4 +2,4 @@ package kr.co.hrmanager.domain.employees;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeesRepository extends JpaRepository<Employees, Integer>  { }
+public interface EmployeesRepository extends JpaRepository<Employees, Long>  { }
