+++ src/main/java/kr/co/hrmanager/domain/jobs/JobsRepository.java
@@ -2,6 +2,6 @@ package kr.co.hrmanager.domain.jobs;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobsRepository extends JpaRepository<Jobs, Integer>  {
+public interface JobsRepository extends JpaRepository<Jobs, Long>  {
 
 }
